+discretizations package: - added method merge to signal_matrix - optimized method DFT - corrected spectral_points to accomodate +controls.setting_tx, +controls.setting_rx - removed spectral_points_base, spectral_points_tx, spectral_points_rx
+math package
- corrected method unique of sequence_increasing
- fixed sequence_increasing_regular to produce zero members because of int64

+physical_values package
- added more is* functions (not tested)
- added more mustBe* functions (not tested)

+controls package:
- optimized constructor for setting
- added method discretize (Fourier transform) to setting; ensures single signal matrices for impulse_responses after transform
- optimized constructor for setting_rx: ensure math.interval, multiple objects / single intervals, ensure correct physical units
- optimized method discretize for setting_rx: use method in superclass
- optimized constructor for setting_tx
- optimized method discretize for setting_tx: use method in superclass for impulse_responses, local transform for excitation_voltages; ensures single signal matrices for excitation_voltages after transform
- optimized method unique for setting_tx: use single signal_matrices to find unique transform coefficients

+transducers package
- optimized array_planar to use concatenation of physical values in centers method

+++ +math/sequence_increasing_regular.m
@@ -54,7 +54,8 @@ classdef sequence_increasing_regular < math.sequence_increasing
             %--------------------------------------------------------------
             members = cell( size( lbs_q ) );
             for index_object = 1:numel( lbs_q )
-                members{ index_object } = ( lbs_q( index_object ):ubs_q( index_object ) ) * deltas( index_object );
+                % use double function to prevent zeros caused by int64
+                members{ index_object } = double( lbs_q( index_object ):ubs_q( index_object ) ) * deltas( index_object );
             end
 
             %--------------------------------------------------------------
